% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fun.R
\name{base_fun}
\alias{base_fun}
\title{Evaluate a Spline Basis Function}
\usage{
base_fun(x, knots, m, j)
}
\arguments{
\item{x}{Numeric vector of data points.}

\item{knots}{List with knot positions and spline settings.}

\item{m}{Number of inner knots.}

\item{j}{Index of the basis function to evaluate.}
}
\value{
Numeric vector with evaluated basis function values.
}
\description{
Computes a single spline basis function at given x-values.
}
