% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mu.R
\name{calc_mu}
\alias{calc_mu}
\title{Estimate Mean Function}
\usage{
calc_mu(
  X,
  y,
  K_mu,
  mu_init,
  sigma_hat,
  lambda_grid,
  lambda_mu = 1,
  max_iterations_mu,
  tolerance
)
}
\arguments{
\item{X}{Design matrix for the mean.}

\item{y}{Response vector.}

\item{K_mu}{Penalty matrix for the mean.}

\item{mu_init}{Initial estimates of the mean.}

\item{sigma_hat}{Current variance estimates.}

\item{lambda_grid}{Grid of smoothing parameters to test.}

\item{lambda_mu}{Starting lambda value.}

\item{max_iterations_mu}{Maximum IRLS iterations.}

\item{tolerance}{Convergence threshold.}
}
\value{
List with updated mean estimates and diagnostics.
}
\description{
Updates the mean curve using penalized IRLS.
}
