% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knots.R
\name{create_knots}
\alias{create_knots}
\title{Create Spline Knot Sequence}
\usage{
create_knots(x, a, b, m = 20, l, equi = TRUE, tp = FALSE, buffer = 0.05)
}
\arguments{
\item{x}{Vector of data points.}

\item{a}{Lower boundary.}

\item{b}{Upper boundary.}

\item{m}{Number of inner knots.}

\item{l}{Degree of the spline.}

\item{equi}{Logical. Equidistant or quantile-based knots.}

\item{tp}{Logical. Use truncated power basis if TRUE.}

\item{buffer}{Percentage buffer beyond data range.}
}
\value{
List containing knot positions and settings.
}
\description{
Generates inner and boundary knots for spline construction.
}
