% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spline.R
\name{fit_spline}
\alias{fit_spline}
\title{Build Spline Design Matrix}
\usage{
fit_spline(x, knots)
}
\arguments{
\item{x}{Data vector.}

\item{knots}{Knot list as created by \code{create_knots}.}
}
\value{
A matrix of spline basis values.
}
\description{
Constructs the spline basis matrix for data points.
}
