% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_values.R
\name{get_initial_values}
\alias{get_initial_values}
\title{Get Initial Estimates for Optimization}
\usage{
get_initial_values(X, Z, y)
}
\arguments{
\item{X}{Design matrix for the mean.}

\item{Z}{Design matrix for the variance.}

\item{y}{Response vector.}
}
\value{
List with initial mean and variance estimates.
}
\description{
Computes starting values for mean and variance curves.
}
