% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pen_mat.R
\name{get_pen_mat}
\alias{get_pen_mat}
\title{Create Penalty Matrix}
\usage{
get_pen_mat(knots, r = 2)
}
\arguments{
\item{knots}{Knot list containing spline info.}

\item{r}{Order of differences for the penalty.}
}
\value{
Penalty matrix.
}
\description{
Constructs the penalty matrix for spline coefficients.
}
