% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_func.R
\name{run_all_funcs}
\alias{run_all_funcs}
\title{Run Full Spline Estimation Pipeline}
\usage{
run_all_funcs(x, y, a, b, m = 20, l, equi = T, tp = F, buffer = 0.05)
}
\arguments{
\item{x}{Predictor values.}

\item{y}{Response values.}

\item{a}{Lower boundary for knots.}

\item{b}{Upper boundary for knots.}

\item{m}{Number of inner knots.}

\item{l}{Degree of splines.}

\item{equi}{Logical. Equidistant or quantile-based knots.}

\item{tp}{Logical. Use truncated power basis if TRUE.}

\item{buffer}{Percentage buffer beyond data range.}
}
\value{
List of fitted model results.
}
\description{
Complete procedure fitting mean and variance splines for data.
}
