% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Iteratively Update Spline Parameters}
\usage{
update_parameters(
  X,
  Z,
  y,
  max_iterations,
  K_mu,
  K_sigma,
  max_iterations_mu = 15,
  max_iterations_sigma = 15,
  tolerance = sqrt(.Machine$double.eps),
  from_mu = 0.5,
  to_mu,
  stepsize_mu,
  from_sigma = 0.5,
  to_sigma,
  stepsize_sigma,
  lambda_init_mu = 1,
  lambda_init_sigma = 1
)
}
\arguments{
\item{X}{Design matrix for the mean.}

\item{Z}{Design matrix for the variance.}

\item{y}{Response vector.}

\item{max_iterations}{Maximum number of outer iterations.}

\item{K_mu}{Penalty matrix for the mean.}

\item{K_sigma}{Penalty matrix for the variance.}

\item{max_iterations_mu}{Maximum IRLS iterations for the mean.}

\item{max_iterations_sigma}{Maximum IRLS iterations for the variance.}

\item{tolerance}{Convergence threshold.}

\item{from_mu}{Start of lambda grid for the mean.}

\item{to_mu}{End of lambda grid for the mean.}

\item{stepsize_mu}{Step size for lambda grid for the mean.}

\item{from_sigma}{Start of lambda grid for the variance.}

\item{to_sigma}{End of lambda grid for the variance.}

\item{stepsize_sigma}{Step size for lambda grid for the variance.}

\item{lambda_init_mu}{Initial lambda value for mean estimation.}

\item{lambda_init_sigma}{Initial lambda value for variance estimation.}
}
\value{
List of fitted curves and diagnostics.
}
\description{
Alternates estimation of mean and variance curves until convergence.
}
